requirements.txt

pytest==3.7.2
pytest-selenium==1.13.0
webdriver-manager==1.7